Reorganize profiling/processing-related code
